split different scatter models into separate files
