+++ src/echosms/__init__.py
@@ -1,7 +1,8 @@
 """Setup the public API for echoSMs."""
+from .utils import Utils
 from .benchmarkdata import BenchMarkData
 from .referencemodels import ReferenceModels
-from .scattermodels import MSSModel, PSMSModel
-from .utils import Utils
+from .mssmodel import MSSModel
+from .psmsmodel import PSMSModel
 
 __all__ = ['BenchMarkData', 'ReferenceModels', 'MSSModel', 'PSMSModel', 'Utils']
